+++ NEWS.md
@@ -4,6 +4,7 @@ ctmm 1.2.1 (2023-11-16)
 * periodic mean summary now uses beta CIs
 * new mean.ctmm() argument formula for functional response estimation
 * new grid argument dr.fn in akde(), occurrence(), pkde()
+* as.telemetry() now imports ATLAS error ellipses
 * new ctmm.boot() argument clamp
 * mean() of occurrence() now time weighted by default
 * new plot method: plot.ctmm()